refactor : 메소드들 숨 좀 쉬게 띄우기
