+++ src/main/java/roomescape/reservation/dto/AddReservationRequest.java
@@ -1,11 +1,11 @@
 package roomescape.reservation.dto;
 
-public class addReservationRequest {
+public class AddReservationRequest {
     private String name;
     private String date;
     private String time;
 
-    public addReservationRequest(String name, String date, String time) {
+    public AddReservationRequest(String name, String date, String time) {
         this.name = name;
         this.date = date;
         this.time = time;
